big update for automatic chia conversion in rpc

+++ CHIA-RPC/General_NS/GlobalVar.cs
@@ -2,6 +2,6 @@
 {
     public static class GlobalVar
     {
-        public static ulong OneChiaInMojos = 1000000000000;
+        public static decimal OneChiaInMojos = 1000000000000;
     }
 }
